+++ CHANGELOG.md
@@ -7,6 +7,7 @@ All notable changes to this project will be documented in this file.
 ## [1.2.0](https://www.nuget.org/packages/KickLib/1.2.0) - [diff](https://github.com/Bukk94/KickLib/compare/v1.1.1...v1.2.0)
 ### Added
 - Support for the `v1/livestreams` API.
+- Support for `UserIdentity` (badges) in `ChatMessageSentEvent` webhook payload
 - Extended subscription webhook payload events with `ExpiresAt`
 - Added `CancellationToken` support to all API methods.